+++ Sources/SwagGenKit/CodeFormatter.swift
@@ -180,7 +180,7 @@ public class CodeFormatter {
 
         return context
     }
-
+    
     func getOperationContext(_ operation: Swagger.Operation) -> Context {
         var context: Context = [:]
 
@@ -200,6 +200,7 @@ public class CodeFormatter {
         context["method"] = operation.method.rawValue.uppercased()
         context["path"] = operation.path
         context["description"] = operation.description
+        context["summary"] = operation.summary
         context["tag"] = operation.tags.first
         context["tags"] = operation.tags